# hardware/server/openapi.yaml
...
openapi: 3.1.0
info:
  title: SitSense API (mock)
  description: 'Raspberry Pi 서버 → 안드로이드 앱. 같은 Wi-Fi에서 `http://<Pi IP>:8000`.


    지금은 **mock 서버**: 시나리오 더미 데이터와 가짜 판단 결과를 준다. 시각은 time.time() 초, 거리 mm, 각도 °, 비율
    0~1. 표시용 글자는 앱이 만든다. 영상 이미지는 어떤 응답에도 없다.'
  version: 0.1.0-mock
paths:
...
  /current:
    get:
      tags:
      - 자세
      summary: 지금 자세 (1~3초마다 호출)
      operationId: current_current_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Current'
        '404':
          description: 아직 기록 없음
...
  /seating/segments:
    get:
      tags:
      - 착석
      summary: 끝난 착석 구간 목록
      operationId: seating_segments_seating_segments_get
      parameters:
      - name: hours
        in: query
        required: false
        schema:
          type: integer
          maximum: 48
          minimum: 1
          description: 최근 몇 시간 안에 끝난 구간
          default: 24
          title: Hours
        description: 최근 몇 시간 안에 끝난 구간
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SeatingSegments'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /calibrate:
    post:
      tags:
      - 기준 자세
      summary: 기준 자세 측정 (앉으면 10초). 세션 없이(온보딩) 또는 세션 중(설정의 다시 측정)
      operationId: calibrate_calibrate_post
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CalibrateResult'
        '409':
          description: 지금 상태에서 할 수 없음 (detail.code 참고)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: 잘못된 값 (detail.code 참고)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /time/sync:
    post:
      tags:
      - 시각
      summary: 앱 시각·시간대 보내기 (연결할 때, 세션 시작 전마다)
      operationId: time_sync_time_sync_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TimeSync'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TimeStatus'
        '409':
          description: 지금 상태에서 할 수 없음 (detail.code 참고)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: 잘못된 값 (detail.code 참고)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /session/start:
    post:
      tags:
      - 세션
      summary: 측정 시작 (시각 동기화 후에만). 기준 자세 측정부터
      operationId: session_start_session_start_post
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionStatus'
        '409':
          description: 지금 상태에서 할 수 없음 (detail.code 참고)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: 잘못된 값 (detail.code 참고)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /session/pause:
    post:
      tags:
      - 세션
      summary: '일시정지: 기록·판단 멈춤, 카메라·거리 끔 (착석 구간은 여기서 끝)'
      operationId: session_pause_session_pause_post
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionStatus'
        '409':
          description: 지금 상태에서 할 수 없음 (detail.code 참고)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: 잘못된 값 (detail.code 참고)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /session/resume:
    post:
      tags:
      - 세션
      summary: 재개
      operationId: session_resume_session_resume_post
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionStatus'
        '409':
          description: 지금 상태에서 할 수 없음 (detail.code 참고)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: 잘못된 값 (detail.code 참고)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /session/stop:
    post:
      tags:
      - 세션
      summary: '정지: 세션 끝, 센서 모두 끔 (앉아도 자동으로 다시 시작하지 않음)'
      operationId: session_stop_session_stop_post
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionStatus'
        '409':
          description: 지금 상태에서 할 수 없음 (detail.code 참고)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: 잘못된 값 (detail.code 참고)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    BaselinePhase:
      type: string
      enum:
      - none
      - waiting_seat
      - measuring
      - ready
      - failed
      title: BaselinePhase
...
    CalibrateResult:
      properties:
        started:
          type: boolean
          title: Started
        seconds:
          type: number
          title: Seconds
          description: 측정에 걸리는 초 (이 동안 바른 자세로 앉기)
      type: object
      required:
      - started
      - seconds
      title: CalibrateResult
    Confidence:
      properties:
        seat:
          type: number
          maximum: 1.0
          minimum: 0.0
          title: Seat
          description: 좌면 판단 확신도 (0~1)
        head:
          type: number
          maximum: 1.0
          minimum: 0.0
          title: Head
          description: 목 판단 확신도 (0~1). 카메라 미인식이면 0
        tilt:
          type: number
          maximum: 1.0
          minimum: 0.0
          title: Tilt
          description: 기울기 판단 확신도 (0~1). 카메라 미인식이면 0
      type: object
      required:
      - seat
      - head
      - tilt
      title: Confidence
    Cues:
      properties:
        neck_drop:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Neck Drop
        face_grow:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Face Grow
        closer:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Closer
        shoulder_angle:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Shoulder Angle
        head_angle:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Head Angle
        head_offset:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Head Offset
      type: object
      title: Cues
      description: 각 단서가 기준값(threshold)을 넘었는지. 판단할 수 없으면 null
    Current:
      properties:
        ts:
          type: number
          title: Ts
          description: 측정 시각 (time.time() 초)
        seated:
          type: boolean
          title: Seated
          description: 앉아 있는지 (seat_state != empty)
        sitting_since:
          anyOf:
          - type: number
          - type: 'null'
          title: Sitting Since
          description: 이번 연속 착석 시작 시각. 자리 비움이면 null
        state_since:
          type: number
          title: State Since
          description: 지금 자세(좌면·목·기울기)가 시작된 시각
        seat_state:
          $ref: '#/components/schemas/SeatState'
        head_state:
          $ref: '#/components/schemas/HeadState'
        tilt_state:
          $ref: '#/components/schemas/TiltState'
        confidence:
          $ref: '#/components/schemas/Confidence'
        postures:
          items:
            type: string
          type: array
          title: Postures
          description: 지금 나타난 나쁜 자세 (cross_left, cross_right, lean_left, lean_right,
            forward_head, tilt_left, tilt_right). 없으면 []
        pending:
          items:
            $ref: '#/components/schemas/Pending'
          type: array
          title: Pending
          description: 확정 대기 중인 자세 변화
        deltas:
          $ref: '#/components/schemas/Deltas'
        cues:
          $ref: '#/components/schemas/Cues'
        distance_mm:
          anyOf:
          - type: integer
          - type: 'null'
          title: Distance Mm
          description: 화면 거리 (mm). 측정 실패면 null
        distance_level:
          $ref: '#/components/schemas/DistanceLevel'
          description: ok 50cm 이상 / caution 40~50cm / warning 40cm 미만 / unknown
        closer_than_baseline:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Closer Than Baseline
          description: 기준보다 10cm 이상 가까운지. 기준값 없으면 null
        pressure:
          items:
            type: integer
          type: array
          title: Pressure
          description: FSR 채널별 원래 값 (0~1023). 순서 = config.yaml pressure.positions
        pressure_ratio:
          anyOf:
          - items:
              type: number
            type: array
          - type: 'null'
          title: Pressure Ratio
          description: 채널별 압력 비율 (합 1). 자리 비움이면 null
        center_of_pressure:
          anyOf:
          - $ref: '#/components/schemas/Point'
          - type: 'null'
          description: 압력 중심. x 왼쪽 -1 ~ 오른쪽 +1, y 뒤 -1 ~ 앞 +1 (사람 기준). 자리 비움이면 null
        pose_detected:
          type: boolean
          title: Pose Detected
          description: 카메라가 사람을 찾았는지
        camera_age_sec:
          anyOf:
          - type: number
          - type: 'null'
          title: Camera Age Sec
          description: 카메라 결과가 몇 초 전 것인지 (초당 약 5장)
        baseline_ready:
          type: boolean
          title: Baseline Ready
          description: 기준 자세가 있는지. false면 deltas 대부분이 null
      type: object
      required:
      - ts
      - seated
      - state_since
      - seat_state
      - head_state
      - tilt_state
      - confidence
      - postures
      - pending
      - deltas
      - cues
      - distance_level
      - pressure
      - pose_detected
      - baseline_ready
      title: Current
      description: 지금 자세 — 홈 화면용, 1~3초마다 호출
    Deltas:
      properties:
        neck_drop:
          anyOf:
          - type: number
          - type: 'null'
          title: Neck Drop
          description: 목 높이가 기준보다 줄어든 비율. +0.12 이상이면 거북목 단서
        face_grow:
          anyOf:
          - type: number
          - type: 'null'
          title: Face Grow
          description: 얼굴 크기가 기준보다 커진 비율. +0.08 이상이면 거북목 단서
        closer_mm:
          anyOf:
          - type: number
          - type: 'null'
          title: Closer Mm
          description: 기준보다 화면에 가까워진 거리 (mm). +면 가까워짐
        shoulder_angle:
          anyOf:
          - type: number
          - type: 'null'
          title: Shoulder Angle
          description: 어깨 선 기울기 변화 (°). + 사람 왼쪽 / - 오른쪽
        head_angle:
          anyOf:
          - type: number
          - type: 'null'
          title: Head Angle
          description: 눈 선 기울기 변화 (°). + 왼쪽 / - 오른쪽
        head_offset:
          anyOf:
          - type: number
          - type: 'null'
          title: Head Offset
          description: 코의 좌우 이동 (어깨 너비 대비 비율). + 왼쪽 / - 오른쪽
      type: object
      title: Deltas
      description: 기준 자세 대비 변화량. 카메라 미인식·기준값 없음이면 null
    DistanceLevel:
      type: string
      enum:
      - ok
      - caution
      - warning
      - unknown
      title: DistanceLevel
      description: 화면 거리 단계 (기준값은 config.yaml의 thresholds)
    EndReason:
      type: string
      enum:
      - away
      - stop
      - pause
      title: EndReason
    ErrorDetail:
      properties:
        code:
          type: string
          title: Code
          description: 앱이 처리할 오류 코드 (CLOCK_NOT_SYNCED, CLOCK_BACKWARD, CLOCK_JUMP_IN_SESSION,
            CLOCK_BEFORE_LAST_RECORD, SESSION_ACTIVE, NOT_RUNNING, NOT_PAUSED, NO_SESSION,
            CONFIRM_REQUIRED, INVALID_TIME, INVALID_TIMEZONE)
        message:
          type: string
          title: Message
          description: 개발자용 설명 (앱 화면 문구는 앱이 만든다)
        resync_required:
          type: boolean
          title: Resync Required
          description: true면 POST /time/sync를 다시 보내야 함
        boot_id:
          type: string
          title: Boot Id
          description: 서버가 켜질 때마다 바뀌는 값. 앱이 기억한 값과 다르면 재시작된 것
        last_record_at:
          anyOf:
          - type: number
          - type: 'null'
          title: Last Record At
          description: CLOCK_BEFORE_LAST_RECORD일 때 Pi의 마지막 기록 시각 (앱이 날짜를 보여줄 때)
      type: object
      required:
      - code
      - message
      - resync_required
      - boot_id
      title: ErrorDetail
    ErrorResponse:
      properties:
        detail:
          $ref: '#/components/schemas/ErrorDetail'
      type: object
      required:
      - detail
      title: ErrorResponse
      description: 오류 응답 (409 상태 충돌, 422 잘못된 값)
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    HeadState:
      type: string
      enum:
      - normal
      - forward
      - unknown
      title: HeadState
      description: 목 전방 자세 (카메라 + 거리센서)
...
    Pending:
      properties:
        kind:
          $ref: '#/components/schemas/StateKind'
          description: 어떤 판단인지 (seat / head / tilt)
        state:
          type: string
          title: State
          description: '바뀔 상태 값 (예: forward)'
        remaining_sec:
          type: number
          minimum: 0.0
          title: Remaining Sec
          description: 확정까지 남은 초
      type: object
      required:
      - kind
      - state
      - remaining_sec
      title: Pending
      description: 새 자세가 나타났지만 아직 short_filter_sec(3초)가 안 지나 확정되지 않은 것
    Point:
      properties:
        x:
          type: number
          title: X
        y:
          type: number
          title: Y
      type: object
      required:
      - x
      - y
      title: Point
...
    SeatState:
      type: string
      enum:
      - normal
      - lean_left
      - lean_right
      - cross_left
      - cross_right
      - empty
      - unknown
      title: SeatState
      description: 좌면 자세 (압력센서)
...
    SeatingSegment:
      properties:
        start:
          type: number
          title: Start
          description: 착석 시작 시각 (압력이 처음 들어온 순간)
        end:
          anyOf:
          - type: number
          - type: 'null'
          title: End
          description: 착석 끝 시각. 진행 중이면 null
        end_reason:
          anyOf:
          - $ref: '#/components/schemas/EndReason'
          - type: 'null'
          description: away 자리 비움 5초 / stop 정지 버튼. 진행 중이면 null
        duration_sec:
          anyOf:
          - type: number
          - type: 'null'
          title: Duration Sec
          description: 구간 길이 (진행 중이면 지금까지)
        first_abnormal_sec:
          anyOf:
          - type: number
          - type: 'null'
          title: First Abnormal Sec
          description: 구간 시작 → 첫 비정상 자세 확정(3초 유지)까지 초. 아직 없으면 null
        normal_sec:
          type: number
          title: Normal Sec
          description: 정상 시간 (목 normal + 기울기 none + 좌면 normal)
        abnormal_sec:
          type: number
          title: Abnormal Sec
          description: 비정상 시간 (나쁜 자세가 하나라도 확정)
        unknown_sec:
          type: number
          title: Unknown Sec
          description: 판단 못 한 시간 (카메라 미인식·켜는 중, 기준 측정 전, 5초 미만 자리 비움)
      type: object
      required:
      - start
      - normal_sec
      - abnormal_sec
      - unknown_sec
      title: SeatingSegment
      description: 착석 구간 한 개. 자리 비움이 5초 이어지면 끝나고, 끝 시각 = 판단 시각 − 5초
    SeatingSegments:
      properties:
        hours:
          type: integer
          title: Hours
        items:
          items:
            $ref: '#/components/schemas/SeatingSegment'
          type: array
          title: Items
      type: object
      required:
      - hours
      - items
      title: SeatingSegments
      description: 끝난 착석 구간 목록 (오래된 것부터). 지금은 서버 메모리에만 있음
    SensorPower:
      type: string
      enum:
      - 'on'
      - 'off'
      - warming
      title: SensorPower
...
    SessionBaseline:
      properties:
        state:
          $ref: '#/components/schemas/BaselinePhase'
          description: none 기준 없음 / waiting_seat 앉기를 기다림 / measuring 측정 중 / ready
            기준 있음 / failed 마지막 측정 실패
        kind:
          anyOf:
          - type: string
          - type: 'null'
          title: Kind
          description: 측정 중이거나 마지막 측정의 종류 (initial / session / recalibration)
        remaining_sec:
          anyOf:
          - type: number
          - type: 'null'
          title: Remaining Sec
          description: 측정 중이면 남은 초 (앉기를 기다리는 중이면 측정 시간 전체)
        fail_reason:
          anyOf:
          - type: string
          - type: 'null'
          title: Fail Reason
          description: 실패 이유 (no_person / sensor_lost / too_much_motion)
        using_previous:
          type: boolean
          title: Using Previous
          description: 이번 측정이 실패해서 직전 성공 기준으로 판단 중인지
        measured_at:
          anyOf:
          - type: number
          - type: 'null'
          title: Measured At
          description: 지금 쓰는 기준을 잰 시각
      type: object
      required:
      - state
      - using_previous
      title: SessionBaseline
      description: 세션의 기준 자세 — 시작 버튼을 누르면 앉은 게 확인된 뒤 10초 측정 (재개 때는 다시 재지 않음)
    SessionSensors:
      properties:
        pressure:
          $ref: '#/components/schemas/SensorPower'
        camera:
          $ref: '#/components/schemas/SensorPower'
          description: warming = 켰지만 아직 사람을 찾기 전 (약 2초)
        distance:
          $ref: '#/components/schemas/SensorPower'
      type: object
      required:
      - pressure
      - camera
      - distance
      title: SessionSensors
      description: 센서 전원 (자리 비움 5초면 카메라·거리 끔, 일시정지면 압력만, 정지면 모두 끔)
    SessionState:
      type: string
      enum:
      - idle
      - running
      - paused
      title: SessionState
    SessionStatus:
      properties:
        state:
          $ref: '#/components/schemas/SessionState'
          description: idle 세션 없음 / running 측정 중 / paused 일시정지
        session_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Session Id
          description: 마지막(또는 지금) 세션 ID. 한 번도 안 했으면 null
        started_at:
          anyOf:
          - type: number
          - type: 'null'
          title: Started At
        ended_at:
          anyOf:
          - type: number
          - type: 'null'
          title: Ended At
          description: 정지한 시각. 진행 중이면 null
        elapsed_sec:
          anyOf:
          - type: number
          - type: 'null'
          title: Elapsed Sec
          description: 경과 시간 (일시정지 시간 제외)
        paused_sec:
          anyOf:
          - type: number
          - type: 'null'
          title: Paused Sec
          description: 일시정지한 시간 합
        seated_sec:
          anyOf:
          - type: number
          - type: 'null'
          title: Seated Sec
          description: 이 세션의 착석 시간 (착석 구간 합, 일시정지 제외)
        seated_now:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Seated Now
          description: 지금 압력이 있는지 (일시정지 중에도). 압력센서가 꺼져 있으면 null
        sensors:
          $ref: '#/components/schemas/SessionSensors'
        clock_synced:
          type: boolean
          title: Clock Synced
        boot_id:
          type: string
          title: Boot Id
        end_reason:
          anyOf:
          - type: string
          - type: 'null'
          title: End Reason
          description: 마지막 세션이 끝난 이유 (stop / baseline_failed). 진행 중이면 null
        baseline:
          $ref: '#/components/schemas/SessionBaseline'
      type: object
      required:
      - state
      - sensors
      - clock_synced
      - boot_id
      - baseline
      title: SessionStatus
      description: 측정 세션 상태 — 홈 하단 측정 바
    StateKind:
      type: string
      enum:
      - seat
      - head
      - tilt
      title: StateKind
...
    TiltState:
      type: string
      enum:
      - none
      - left
      - right
      - unknown
      title: TiltState
      description: 상체 좌우 기울어짐 (카메라). 방향은 사람 기준
    TimeStatus:
      properties:
        synced:
          type: boolean
          title: Synced
          description: 이 서버가 켜진 뒤 동기화했는지
        boot_id:
          type: string
          title: Boot Id
          description: 서버가 켜질 때마다(재부팅 포함) 바뀜. 바뀌었으면 다시 동기화
        timezone:
          anyOf:
          - type: string
          - type: 'null'
          title: Timezone
        app_time:
          anyOf:
          - type: number
          - type: 'null'
          title: App Time
          description: 앱 기준 지금 시각. 동기화 전이면 null
        local_time:
          anyOf:
          - type: string
          - type: 'null'
          title: Local Time
          description: app_time을 시간대로 바꾼 ISO 문자열 (확인용)
        synced_at:
          anyOf:
          - type: number
          - type: 'null'
          title: Synced At
          description: 마지막 동기화 때 앱이 보낸 시각
        resync_required:
          type: boolean
          title: Resync Required
          description: true면 POST /time/sync 필요
        offset_change_sec:
          anyOf:
          - type: number
          - type: 'null'
          title: Offset Change Sec
          description: 다시 동기화했을 때 시각이 바뀐 양 (초). 첫 동기화·조회에서는 null
        applied:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Applied
          description: 이번 동기화를 적용했는지. 0.5초 이내로 뒤로 가는 변경은 false(시각 그대로). 조회에서는 null
        message:
          anyOf:
          - type: string
          - type: 'null'
          title: Message
      type: object
      required:
      - synced
      - boot_id
      - resync_required
      title: TimeStatus
      description: Pi의 시각 동기화 상태. Pi OS 시계는 바꾸지 않고 '앱 시각 − Pi 단조시계' 차이만 저장
    TimeSync:
      properties:
        app_time:
          type: number
          title: App Time
          description: 앱의 현재 시각 (time.time()과 같은 초, 소수 가능)
          examples:
          - 1791207600.123
        timezone:
          type: string
          title: Timezone
          description: 시간대 (Asia/Seoul만 가능)
          default: Asia/Seoul
          examples:
          - Asia/Seoul
      type: object
      required:
      - app_time
      title: TimeSync
      description: '앱 → Pi: 현재 시각과 시간대. 연결할 때와 세션 시작 전마다 보낸다'
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError